GetModbusSlaveTable moved from abstract class to implementation class.
