+++ io.openems.edge.ess.generic/src/io/openems/edge/ess/generic/offgrid/GenericOffGridEssImpl.java
@@ -25,7 +25,9 @@ import io.openems.edge.common.component.ComponentManager;
 import io.openems.edge.common.component.OpenemsComponent;
 import io.openems.edge.common.event.EdgeEventConstants;
 import io.openems.edge.common.modbusslave.ModbusSlave;
+import io.openems.edge.common.modbusslave.ModbusSlaveNatureTable;
 import io.openems.edge.common.modbusslave.ModbusSlaveTable;
+import io.openems.edge.common.modbusslave.ModbusType;
 import io.openems.edge.common.startstop.StartStop;
 import io.openems.edge.common.startstop.StartStoppable;
 import io.openems.edge.ess.api.HybridEss;
@@ -149,15 +151,6 @@ public class GenericOffGridEssImpl
 	}
 
 	@Override
-	public ModbusSlaveTable getModbusSlaveTable(AccessMode accessMode) {
-		return new ModbusSlaveTable(//
-				OpenemsComponent.getModbusSlaveNatureTable(accessMode), //
-				SymmetricEss.getModbusSlaveNatureTable(accessMode), //
-				ManagedSymmetricEss.getModbusSlaveNatureTable(accessMode) //
-		);
-	}
-
-	@Override
 	protected ChannelManager getChannelManager() {
 		return this.channelManager;
 	}
@@ -192,5 +185,17 @@ public class GenericOffGridEssImpl
 			// Set only if value changed
 			this.stateMachine.forceNextState(OffGridState.UNDEFINED);
 		}
+	}
+	
+	@Override
+	public ModbusSlaveTable getModbusSlaveTable(AccessMode accessMode) {
+		return new ModbusSlaveTable(//
+				OpenemsComponent.getModbusSlaveNatureTable(accessMode), //
+				SymmetricEss.getModbusSlaveNatureTable(accessMode), //
+				ManagedSymmetricEss.getModbusSlaveNatureTable(accessMode), //
+				StartStoppable.getModbusSlaveNatureTable(accessMode), //
+				ModbusSlaveNatureTable.of(GenericOffGridEss.class, accessMode, 100) //
+						.channel(0, GenericOffGridEss.ChannelId.STATE_MACHINE, ModbusType.UINT16) //
+						.build());
 	}
 }
